refactor: decompose settings
